requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-Mail==0.10.0
python-dotenv==1.0.0